Added GLM and InputManager
Forgot to added GLM in the initial commit, whoops. I made it so you can
close the window now too.

+++ MegaEngine/CoreEngine.h
@@ -24,7 +24,7 @@ public:
 
 private:
 	bool m_running = false;
-	float m_frameTime = 60.0f;
+	double m_frameTime = 60.0;
 
 	Viewport* m_viewport = nullptr;
 	SceneManager* m_sceneManager = nullptr;
